Created a seperate module file for the userinputs
